+++ CHANGELOG.md
@@ -10,6 +10,8 @@ that you can set version constraints properly.
 
 #### [Unreleased] - now
 
+* `Added`: validation of targets against feature target_type contract
+  ([#78](https://github.com/codebreakdown/togls/issues/78))
 * `Added`: setting a default feature target type
   ([#69](https://github.com/codebreakdown/togls/issues/69))
 * `Added`: logging for target type mismatches/erroneous states